requirements_frozen.txt
aniso8601==8.0.0
attrs==19.3.0
Click==7.0
dill==0.3.1.1
Flask==1.1.1
Flask-RESTful==0.3.7
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
jsonref==0.2
markdown2==2.3.8
MarkupSafe==1.1.1
mllaunchpad==0.0.7
numpy==1.18.1
pandas==0.25.3
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
ramlfications==0.1.9
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
termcolor==1.1.0
Werkzeug==0.16.0
xmltodict==0.12.0
